Before changing callDiamondProxy()
